Preparing to Implement( New ) ChatBot
The `test` branch does have a ChatBot but it has a lot of issues.
Therefore, I created this branch to restart the implementation of the AI
assistant.

I now where most of the errors are coming from and I am going to try to
fix them before starting to work on the chatbot itself.